jigyasa/gui/requirements.txt
# JIGYASA GUI Requirements
flask>=2.3.0
flask-socketio>=5.3.0
flask-cors>=4.0.0
python-socketio[client]>=5.9.0
gitpython>=3.1.0
beautifulsoup4>=4.12.0
feedparser>=6.0.0
aiohttp>=3.8.0
requests>=2.31.0